% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bioFabric_htmlwidget.R
\name{bioFabric_htmlwidget-shiny}
\alias{bioFabric_htmlwidget-shiny}
\alias{bioFabric_htmlwidgetOutput}
\alias{renderBioFabric_htmlwidget}
\title{Shiny bindings for bioFabric_htmlwidget}
\usage{
bioFabric_htmlwidgetOutput(outputId, width = "100\%", height = "400px")

renderBioFabric_htmlwidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a bioFabric_htmlwidget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using bioFabric_htmlwidget within Shiny
applications and interactive Rmd documents.
}

